requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
click==7.1.2
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
flake8==3.8.3
gunicorn==20.0.4
iniconfig==1.0.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.2.0
